bring back SparsityConfigMetadata
